+++ requirements.txt
@@ -1,0 +1,9 @@
+dotenv==0.9.9
+mysql-connector==2.2.9
+numpy==2.2.4
+pandas==2.2.3
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+pytz==2025.2
+six==1.17.0
+tzdata==2025.2
