+++ scripts/main.py
